feat: allow for custom shortcut paths and dont place shortcut if path is empty
untested lets hope this works

+++ blueprint.sh
@@ -17,6 +17,7 @@ FOLDER=$(realpath "$(dirname "$0" 2> /dev/null)" 2> /dev/null) || FOLDER="$BLUEP
 OWNERSHIP="www-data:www-data" #;
 WEBUSER="www-data" #;
 USERSHELL="/bin/bash" #;
+SHORTCUT_PATH="/usr/local/bin"
 
 # Check if the script is being sourced - and if so - load bash autocompletion.
 if [[ "${BASH_SOURCE[0]}" != "${0}" ]]; then
@@ -209,20 +210,24 @@ assignflags() {
 
 # Adds the "blueprint" command to the /usr/local/bin directory and configures the correct permissions for it.
 placeshortcut() {
-  PRINT INFO "Placing Blueprint command shortcut.."
-
-  rm -f scripts/helpers/blueprint.bak
-  cp "scripts/helpers/blueprint" "scripts/helpers/blueprint.bak"
-  sed -i "s~BLUEPRINT_FOLDER_HERE~$FOLDER~g" "scripts/helpers/blueprint.bak"
-
-  rm -f /usr/local/bin/blueprint
-  mv scripts/helpers/blueprint.bak /usr/local/bin/blueprint
-
-  {
-    chmod 755 \
-      "$FOLDER/blueprint.sh" \
-      /usr/local/bin/blueprint
-  } >> "$BLUEPRINT__DEBUG"
+  if [[ $SHORTCUT_PATH != "" ]]; then
+    PRINT INFO "Placing Blueprint command shortcut.."
+  
+    rm -f scripts/helpers/blueprint.bak
+    cp "scripts/helpers/blueprint" "scripts/helpers/blueprint.bak"
+    sed -i "s~BLUEPRINT_FOLDER_HERE~$FOLDER~g" "scripts/helpers/blueprint.bak"
+  
+    rm -f "$SHORTCUT_PATH/blueprint"
+    mv scripts/helpers/blueprint.bak "$SHORTCUT_PATH/blueprint"
+  
+    {
+      chmod 755 \
+        "$FOLDER/blueprint.sh" \
+        "$SHORTCUT_PATH/blueprint"
+    } >> "$BLUEPRINT__DEBUG"
+  else
+    PRINT DEBUG "Shortcut not placed as \$SHORTCUT_PATH is empty"
+  fi
 }
 if ! [ -x "$(command -v blueprint)" ]; then placeshortcut; fi
